♻️ Merge HttpError into base class
